feat: remove InMemory abstractions from infrastructure layer [BLDX-874]
Reduces over-abstraction by eliminating InMemory classes from the
infrastructure runtime path. Dapr is now required for all modes
(dev, test runner, prod).

Changes:
- Move credential interception from InMemorySecretStore.set() to
  DaprStateStore.save() (state store is writable, secret store isn't)
- credentials/resolver.py: check state store for cred: prefixed keys
  before falling back to secret store and DaprCredentialVault
- main.py: remove InMemory fallback path — require Dapr sidecar
  (clear error message if DAPR_HTTP_PORT not set)
- Delete InMemoryStateStore, InMemorySecretStore, InMemoryPubSub,
  InMemoryCredentialVault, InMemoryBinding from infrastructure/
- Make Mock classes self-contained (implement Protocol directly with
  dict storage + call tracking, no InMemory base class)
- Remove InMemory exports from infrastructure/__init__.py
- Update all test imports to use testing.mocks

InMemory classes are gone from the runtime path. Mock classes in
testing/mocks.py remain for unit tests (they implement the Protocol
directly without InMemory base classes).

Tests: 2072 passing
