+++ application_sdk/infrastructure/credential_vault.py
@@ -72,30 +72,3 @@ class CredentialVault(Protocol):
             CredentialVaultError: If resolution fails.
         """
         ...
-
-
-class InMemoryCredentialVault:
-    """In-memory credential vault for testing.
-
-    WARNING: Do not use in production. Credentials are stored in plain text.
-    """
-
-    def __init__(self, credentials: dict[str, dict[str, Any]] | None = None) -> None:
-        self._credentials: dict[str, dict[str, Any]] = credentials or {}
-
-    async def get_credentials(self, credential_guid: str) -> dict[str, Any]:
-        """Get credentials by GUID."""
-        if credential_guid not in self._credentials:
-            raise CredentialVaultError(
-                f"Credential '{credential_guid}' not found",
-                credential_guid=credential_guid,
-            )
-        return dict(self._credentials[credential_guid])
-
-    def set(self, credential_guid: str, credentials: dict[str, Any]) -> None:
-        """Register credentials under *credential_guid* (for test setup)."""
-        self._credentials[credential_guid] = credentials
-
-    def clear(self) -> None:
-        """Remove all registered credentials (for test teardown)."""
-        self._credentials.clear()
