+++ application_sdk/infrastructure/pubsub.py
@@ -123,104 +123,3 @@ class PubSub(Protocol):
             PubSubError: If subscription fails.
         """
         ...
-
-
-class InMemorySubscription:
-    """In-memory subscription for testing."""
-
-    def __init__(self, topic: str, pubsub: "InMemoryPubSub") -> None:
-        self._topic = topic
-        self._pubsub = pubsub
-        self._active = True
-
-    @property
-    def topic(self) -> str:
-        return self._topic
-
-    @property
-    def is_active(self) -> bool:
-        return self._active
-
-    async def unsubscribe(self) -> None:
-        """Unsubscribe from the topic."""
-        self._active = False
-        self._pubsub._remove_subscription(self._topic, self)
-
-
-class InMemoryPubSub:
-    """In-memory pub/sub for testing.
-
-    Messages are delivered immediately to all subscribers.
-    """
-
-    def __init__(self) -> None:
-        self._subscriptions: dict[
-            str, list[tuple[MessageHandler, InMemorySubscription]]
-        ] = {}
-        self._published: list[Message] = []
-
-    async def publish(
-        self,
-        topic: str,
-        data: dict[str, Any],
-        *,
-        metadata: dict[str, str] | None = None,
-    ) -> None:
-        """Publish a message."""
-        import uuid
-
-        message = Message(
-            data=data,
-            metadata=metadata or {},
-            id=str(uuid.uuid4()),
-            topic=topic,
-        )
-        self._published.append(message)
-
-        # Deliver to subscribers
-        if topic in self._subscriptions:
-            for handler, subscription in self._subscriptions[topic]:
-                if subscription.is_active:
-                    await handler(message)
-
-    async def subscribe(
-        self,
-        topic: str,
-        handler: MessageHandler,
-    ) -> Subscription:
-        """Subscribe to a topic."""
-        if topic not in self._subscriptions:
-            self._subscriptions[topic] = []
-
-        subscription = InMemorySubscription(topic, self)
-        self._subscriptions[topic].append((handler, subscription))
-        return subscription
-
-    def _remove_subscription(
-        self,
-        topic: str,
-        subscription: InMemorySubscription,
-    ) -> None:
-        """Remove a subscription (called by InMemorySubscription)."""
-        if topic in self._subscriptions:
-            self._subscriptions[topic] = [
-                (h, s) for h, s in self._subscriptions[topic] if s != subscription
-            ]
-
-    def get_published(self, topic: str | None = None) -> list[Message]:
-        """Get published messages (for testing).
-
-        Args:
-            topic: Optional topic filter.
-
-        Returns:
-            List of published messages.
-        """
-        if topic:
-            return [m for m in self._published if m.topic == topic]
-        return list(self._published)
-
-    def clear(self) -> None:
-        """Clear all subscriptions and published messages (for testing)."""
-        self._subscriptions.clear()
-        self._published.clear()
